+++ addons/zylann.editor_debugger/plugin.gd
@@ -4,6 +4,7 @@ extends EditorPlugin
 const DockScene = preload("dock.tscn")
 
 const Dock = preload("dock.gd")
+const Util = preload("util.gd")
 
 var _dock: Dock = null
 
@@ -24,5 +25,14 @@ func _exit_tree() -> void:
 
 func _on_EditorDebugger_node_selected(node: Node) -> void:
 	if _dock.is_inspection_enabled():
+		if _is_under_editor_inspector(node):
+			# Prevent inspecting the inspector, because unfortunately it can crash.
+			# Something inside Godot is not handling well the possibility that what is inspected
+			# could be freed anytime.
+			return
 		# Oops.
 		EditorInterface.inspect_object(node)
+
+
+func _is_under_editor_inspector(node: Node) -> bool:
+	return Util.get_node_in_parents(node, EditorInspector) != null