+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a more detailed list of past and incoming changes, see the commit history.
 This version requires Godot 4.2 or later.
 
 - Added a menu to copy node access codes using `get_child(index)` instead of names (thanks to ryevdokimov)
+- Disabled inspection of nodes below `EditorInspector` because it can crash Godot.
 - Fixed optional GDScript warnings that could show up when addon warnings are enabled (thanks to pineapplemachine); breaks compatibility with versions of Godot older than 4.2
 - Fixed focus with F12 wasn't always reaching the right node in the tree view (PR #14)